+++ app/sign-in/page.tsx
@@ -3,6 +3,10 @@ import { redirect } from 'next/navigation'
 import { auth } from './../../auth'
 import { LoginButton } from './../../components/login-button'
 
+import SaraPortrait from './../../public/Sara_Cartoon_Portrait.png'
+import PolyverseLogo from './../../public/Polyverse logo medium.jpg'
+import BoostExplanation from 'components/boost-explanation'
+
 export default async function SignInPage() {
   const session = await auth()
   // redirect to home if user is already logged in
@@ -10,8 +14,25 @@ export default async function SignInPage() {
     redirect('/')
   }
   return (
-    <div className="flex h-[calc(100vh-theme(spacing.16))] items-center justify-center py-10">
+    <div className="flex flex-col items-center justify-center p-4 ">
+      <img
+        src={SaraPortrait.src} // Adjust the path to your image
+        alt="Sara's Portrait"
+        width={200} // Adjust the width as needed
+        height={200} // Adjust the height as needed
+      />
+      <p className="font-medium italic text-8xl">Sara</p>
+      <div className="flex items-center justify-center m-5">
+        <p className="font-medium text-5xl mr-3">by</p>
+        <img
+          src={PolyverseLogo.src} // Adjust the path to your image
+          alt="Sara's Portrait"
+          width={200} // Adjust the width as needed
+          height={200} // Adjust the height as needed
+        />
+      </div>
       <LoginButton />
+      <BoostExplanation />
     </div>
   )
 }